\docType{data}
\name{oilflow}
\alias{oilflow}
\title{Oilflow dataset}
\format{A data frame with 1000 observations of 12 variables}
\description{
  Well-known dataset from the paper by Bishop, Christopher
  M., Markus Svensén, and Christopher KI Williams: "GTM:
  The generative topographic mapping." Neural computation
  10.1 (1998): pp. 215-234.
}
\author{
  Bishop, Christopher
}
\keyword{datasets}

