\name{print.gtm}
\alias{print.gtm}
\title{Formatted output of a given Generative Topographic Mapping.}
\usage{
  \method{print}{gtm} (x, ...)
}
\arguments{
  \item{x}{an instance of \code{gtm} (see
  \code{\link{gtm.compute}})}

  \item{...}{additional arguments to be passed on to a
  final print statement}
}
\description{
  Formatted output of a given Generative Topographic
  Mapping.
}
\seealso{
  \code{\link{gtm.compute}}
}

