% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{oilflow}
\alias{oilflow}
\title{Oilflow dataset}
\format{A data frame with 1000 observations of 12 variables}
\description{
Well-known dataset from Bishop et al., "GTM: The generative topographic
mapping.", Neural computation 10.1, pp. 215-234, 1998.
}
\author{
Bishop, Christopher
}
\keyword{datasets}

